refactor: discountpolicy 인터페이스 생성
