+++ src/main/java/christmas/service/EventService.java
@@ -1,6 +1,7 @@
 package christmas.service;
 
 import christmas.domain.event.DiscountPolicy;
+import christmas.domain.event.SimpleDiscountPolicy;
 import christmas.domain.event.Event;
 import christmas.domain.event.EventPlanner;
 import christmas.domain.order.Order;
@@ -8,8 +9,8 @@ import christmas.domain.order.Order;
 public class EventService {
     public Event getEventInfo(Order order) {
 
-        DiscountPolicy discountPolicy = new DiscountPolicy();
-        EventPlanner eventPlanner = new EventPlanner(discountPolicy);
+        DiscountPolicy simpleDiscountPolicy = new SimpleDiscountPolicy();
+        EventPlanner eventPlanner = new EventPlanner(simpleDiscountPolicy);
 
         return eventPlanner.event(order);
     }
